requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
streamlit==1.38.0
tensorflow==2.17.0    # If using TensorFlow
torch==2.4.1          # If using PyTorch
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
Pillow==10.4.0         # If using image processing
python-multipart==0.0.5
